// src/core/integrations/enum/message-type.enum.ts
export enum MessageTypeEnum {
  TEXT = 'text',
  IMAGE = 'image',
  VIDEO = 'video',
  AUDIO = 'audio',
  VOICE = 'voice',
  DOCUMENT = 'document',
  STICKER = 'sticker',
  REEL = 'reel',
  LOCATION = 'location',
  CONTACT = 'contact',
  CONTACT_MULTI = 'contact_multi',
  LIST = 'list',
  LIST_RESPONSE = 'list_response',
  BUTTONS_RESPONSE = 'buttons_response',
  REACTION = 'reaction',
  PAYMENT = 'payment',
  ORDER = 'order',
  POLL = 'poll',
  EVENT = 'event',
  UNKNOWN = 'unknown',
}